# MAEBACKEND/Endpoints/openapi_specification.yaml
openapi: 3.0.3
info:
  title: EasyRent API - Sistema Completo de Marketplace Inmobiliario
  description: |
    API completa para el marketplace inmobiliario EasyRent con todas las funcionalidades:
    - 🔐 **Autenticación**: JWT + API Keys (Compatible con base de datos PostgreSQL)
    - 👥 **Gestión de usuarios y agencias**
    - 🏠 **Listings y multimedia**
    - 🔍 **Búsqueda avanzada y filtros**
    - 💎 **Sistema de planes y facturación**
    - 📊 **Analytics y reportes**
    - ✅ **Verificación y moderación**
    - 🔗 **Integraciones externas**
    
    ## Tecnologías
    - **Framework**: FastAPI + Python 3.11+
    - **Base de datos**: PostgreSQL 17
    - **Cache**: Redis
    - **Message Broker**: Apache Kafka
    - **Autenticación**: JWT (Bearer Token)
    
    ## Notas de Implementación
    - **Autenticación Firebase**: Sistema de autenticación completo con Firebase Authentication
    - **Flujo de autenticación**: Frontend → Firebase → Backend verification → JWT tokens
    - **Sin manejo de passwords**: Toda la gestión de credenciales delegada a Firebase
    - **Soporte híbrido**: Funciona con o sin Firebase configurado (modo desarrollo)
    - **Schemas actualizados**: Los modelos de datos coinciden con la implementación real en FastAPI
    
    ## Autenticación Firebase
    Para autenticarse, el cliente debe:
    1. Autenticarse con Firebase (frontend)
    2. Obtener el Firebase ID token
    3. Enviar el token al endpoint `/auth/login`
...
    
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /agencies:
    get:
      tags: [Agencies]
      summary: Listar agencias
      description: Obtiene una lista de agencias inmobiliarias con filtros opcionales
      operationId: list_agencies
      security: []
      parameters:
        - name: city
          in: query
          description: Filtrar por ciudad
          schema:
            type: string
            example: "Lima"
        - name: verified
          in: query
          description: Solo agencias verificadas
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: Lista de agencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgencyResponse'
        '500':
          description: Error interno del servidor

    post:
      tags: [Agencies]
      summary: Crear agencia
      description: Crea una nueva agencia inmobiliaria
      operationId: create_agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgencyRequest'
      responses:
        '201':
          description: Agencia creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyResponse'
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor
...
  /listings:
    get:
      tags: [Listings]
      summary: Listar propiedades
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: operation_type
          in: query
          schema:
            type: string
            enum: [sale, rent, temp_rent, pre_sale]
        - name: property_type
          in: query
          schema:
            type: string
            enum: [house, apartment, studio, room, office, commercial, land, warehouse, garage, other]
        - name: city
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: min_price
          in: query
          schema:
            type: number
        - name: max_price
          in: query
          schema:
            type: number
        - name: bedrooms
          in: query
          schema:
            type: integer
        - name: bathrooms
          in: query
          schema:
            type: integer
        - name: min_area
          in: query
          schema:
            type: number
        - name: max_area
          in: query
          schema:
            type: number
        - name: min_age_years
          in: query
          schema:
            type: integer
            minimum: 0
            description: Antigüedad mínima de la vivienda en años
        - name: max_age_years
          in: query
          schema:
            type: integer
            minimum: 0
            description: Antigüedad máxima de la vivienda en años
        - name: features
          in: query
          schema:
            type: array
            items:
              type: string
        - name: amenities
          in: query
          schema:
            type: array
            items:
              type: string
        - name: verified
          in: query
          schema:
            type: boolean
        - name: pet_friendly
          in: query
          schema:
            type: boolean
            description: Filtrar por propiedades que permiten mascotas
        - name: furnished
          in: query
          schema:
            type: boolean
            description: Filtrar por propiedades amuebladas/amobladas
        - name: rental_mode
          in: query
          schema:
            type: string
            enum: [full_property, private_room, shared_room]
            description: Modalidad de alquiler
        - name: airbnb_eligible
          in: query
          schema:
            type: boolean
            description: Solo propiedades elegibles para Airbnb
        - name: min_airbnb_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
            description: Score mínimo de elegibilidad Airbnb (0-100)
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, area_asc, area_desc, date_asc, date_desc, age_asc, age_desc, relevance]
      responses:
        '200':
          description: Lista de propiedades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedListings'

    post:
      tags: [Listings]
      summary: Crear nueva propiedad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Propiedad creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
...
  /search:
    get:
      tags: [Search]
      summary: Búsqueda general
      security: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
            description: Texto de búsqueda
        - name: location
          in: query
          schema:
            type: string
            description: Ubicación (ciudad, distrito, etc.)
        - name: lat
          in: query
          schema:
            type: number
            description: Latitud para búsqueda por proximidad
        - name: lng
          in: query
          schema:
            type: number
            description: Longitud para búsqueda por proximidad
        - name: radius
          in: query
          schema:
            type: number
            description: Radio en kilómetros
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
...
  /subscriptions:
    get:
      tags: [Plans]
      summary: Obtener suscripciones del usuario
      responses:
        '200':
          description: Suscripciones del usuario

    post:
      tags: [Plans]
      summary: Crear suscripción
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id]
              properties:
                plan_id:
                  type: string
                  format: uuid
                payment_method_id:
                  type: string
                billing_cycle:
                  type: string
                  enum: [monthly, yearly]
      responses:
        '201':
          description: Suscripción creada
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Obtener notificaciones
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [new_listing, price_change, lead, verification, payment]
        - name: read
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Lista de notificaciones
...
  /admin/listings:
    get:
      tags: [Admin]
      summary: Gestión de listings (admin)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, inactive, sold, rented, expired, flagged]
        - name: verification_status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        '200':
          description: Lista de listings para administración
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
        
        **Ejemplo**: `Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
  parameters:
    PageParam:
      name: page
      in: query
      description: Número de página (empezando en 1)
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Número de elementos por página
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
  schemas:
...
    CreateAgencyRequest:
      type: object
      required: [name, phone, email, address, city]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: Nombre de la agencia inmobiliaria
          example: "Inmobiliaria Lima Norte"
        phone:
          type: string
          minLength: 6
          maxLength: 20
          description: Teléfono de contacto
          example: "+51987654321"
        email:
          type: string
          format: email
          description: Email de contacto
          example: "contacto@inmobiliarialima.com"
        address:
          type: string
          minLength: 2
          maxLength: 255
          description: Dirección física de la agencia
          example: "Av. La Marina 123, San Miguel"
        city:
          type: string
          minLength: 2
          maxLength: 100
          description: Ciudad donde opera
          example: "Lima"
        district:
          type: string
          maxLength: 100
          nullable: true
          description: Distrito específico
          example: "San Miguel"
...
    AgencyResponse:
      type: object
      properties:
        id:
          type: string
          description: ID único de la agencia (UUID como string)
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Nombre de la agencia
          example: "Inmobiliaria Lima Norte"
        phone:
          type: string
          description: Teléfono de contacto
          example: "+51987654321"
        email:
          type: string
          format: email
          description: Email de contacto
          example: "contacto@inmobiliarialima.com"
        address:
          type: string
          description: Dirección física
          example: "Av. La Marina 123, San Miguel"
        city:
          type: string
          description: Ciudad
          example: "Lima"
        district:
          type: string
          nullable: true
          description: Distrito
          example: "San Miguel"
        status:
          type: string
          enum: [active, inactive, suspended, pending_verification, verified]
          description: Estado de la agencia
          example: "active"
        verified:
          type: boolean
          description: Si la agencia está verificada
          example: true
        created_at:
          type: string
          format: date-time
          description: Fecha de creación
          example: "2025-08-28T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Fecha de última actualización
          example: "2025-09-06T10:30:00Z"
...
    Listing:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único de la propiedad
          example: "456e7890-e89b-12d3-a456-426614174001"
        title:
          type: string
          description: Título de la propiedad
          example: "Hermoso departamento en Miraflores con vista al mar"
        description:
          type: string
          description: Descripción detallada
          example: "Amplio departamento de 3 dormitorios..."
        operation_type:
          type: string
          enum: [sale, rent, temp_rent, pre_sale]
          description: Tipo de operación
          example: "sale"
        property_type:
          type: string
          enum: [house, apartment, studio, room, office, commercial, land, warehouse, garage, other]
          description: Tipo de propiedad
          example: "apartment"
        furnished:
          type: boolean
          description: Si la propiedad está amueblada/amoblada
          example: true
        rental_mode:
          type: string
          enum: [full_property, private_room, shared_room]
          description: Modalidad de alquiler (solo aplicable para property_type room o studio)
          example: "private_room"
        airbnb_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Score de elegibilidad para Airbnb (0-100)
          example: 85
        airbnb_eligible:
          type: boolean
          description: Si la propiedad es elegible para estilo Airbnb
          example: true
        price:
          type: number
          description: Precio de la propiedad
          example: 250000
        currency:
          type: string
          default: "PEN"
          description: Moneda del precio
          example: "USD"
        area:
          type: number
          description: Área en metros cuadrados
          example: 120.5
        bedrooms:
          type: integer
          minimum: 0
          description: Número de dormitorios
          example: 3
        bathrooms:
          type: integer
          minimum: 0
          description: Número de baños
          example: 2
        parking_spots:
          type: integer
          minimum: 0
          description: Espacios de estacionamiento
          example: 1
        address:
          type: string
          description: Dirección completa
          example: "Av. Larco 123, Miraflores"
        city:
          type: string
          description: Ciudad
          example: "Lima"
        district:
          type: string
          description: Distrito
          example: "Miraflores"
        latitude:
          type: number
          format: double
          description: Latitud
          example: -12.1191
        longitude:
          type: number
          format: double
          description: Longitud
          example: -77.0372
        features:
          type: array
          items:
            type: string
          description: Características de la propiedad
          example: ["balcón", "cocina_equipada", "aire_acondicionado"]
        amenities:
          type: array
          items:
            type: string
          description: Amenidades del edificio/complejo
          example: ["piscina", "gimnasio", "portería_24h"]
        status:
          type: string
          enum: [draft, active, inactive, sold, rented, expired]
          description: Estado de la propiedad
          example: "active"
        verified:
          type: boolean
          description: Si la propiedad está verificada
          example: true
        views_count:
          type: integer
          description: Número total de visualizaciones
          example: 1245
        leads_count:
          type: integer
          description: Número de leads generados
          example: 15
        images_count:
          type: integer
          description: Número de imágenes
          example: 8
        videos_count:
          type: integer
          description: Número de videos
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2025-08-28T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-08-28T10:00:00Z"
...
    CreateListingRequest:
      type: object
      required: [title, operation_type, property_type, price, area, address, city, district]
      properties:
        title:
          type: string
          minLength: 10
          maxLength: 200
          description: Título descriptivo de la propiedad
          example: "Hermoso departamento en Miraflores con vista al mar"
        description:
          type: string
          maxLength: 2000
          description: Descripción detallada
          example: "Amplio departamento de 3 dormitorios ubicado en la mejor zona de Miraflores..."
        operation_type:
          type: string
          enum: [sale, rent, temp_rent, pre_sale]
          description: Tipo de operación
          example: "sale"
        property_type:
          type: string
          enum: [house, apartment, studio, room, office, commercial, land, warehouse, garage, other]
          description: Tipo de propiedad
          example: "apartment"
        furnished:
          type: boolean
          description: Si la propiedad está amueblada/amoblada
          example: false
        rental_mode:
          type: string
          enum: [full_property, private_room, shared_room]
          description: Modalidad de alquiler (solo aplicable para rooms o studios en renta)
          example: "full_property"
        price:
          type: number
          minimum: 0
          description: Precio de la propiedad
          example: 250000
        currency:
          type: string
          enum: [PEN, USD]
          default: "PEN"
          description: Moneda del precio
          example: "USD"
        area:
          type: number
          minimum: 1
          description: Área en metros cuadrados
          example: 120.5
        bedrooms:
          type: integer
          minimum: 0
          maximum: 20
          description: Número de dormitorios
          example: 3
        bathrooms:
          type: integer
          minimum: 0
          maximum: 20
          description: Número de baños
          example: 2
        parking_spots:
          type: integer
          minimum: 0
          maximum: 10
          description: Espacios de estacionamiento
          example: 1
        address:
          type: string
          minLength: 10
          maxLength: 200
          description: Dirección completa
          example: "Av. Larco 123, Miraflores"
        city:
          type: string
          minLength: 2
          maxLength: 50
          description: Ciudad
          example: "Lima"
        district:
          type: string
          minLength: 2
          maxLength: 50
          description: Distrito
          example: "Miraflores"
        features:
          type: array
          items:
            type: string
          maxItems: 20
          description: Características de la propiedad
          example: ["balcón", "cocina_equipada", "aire_acondicionado"]
        amenities:
          type: array
          items:
            type: string
          maxItems: 20
          description: Amenidades del edificio/complejo
          example: ["piscina", "gimnasio", "portería_24h"]
...
    PaginatedListings:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          description: Página actual
          example: 1
        per_page:
          type: integer
          description: Elementos por página
          example: 20
        total:
          type: integer
          description: Total de elementos
          example: 150
        last_page:
          type: integer
          description: Última página
          example: 8
        from:
          type: integer
          description: Elemento inicial en la página
          example: 1
        to:
          type: integer
          description: Elemento final en la página
          example: 20
        has_next:
          type: boolean
          description: Si hay página siguiente
          example: true
        has_prev:
          type: boolean
          description: Si hay página anterior
          example: false
    SearchResults:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        facets:
          type: object
          description: Facetas para filtros
          properties:
            cities:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: "Lima"
                  count:
                    type: integer
                    example: 45
            districts:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: "Miraflores"
                  city:
                    type: string
                    example: "Lima"
                  count:
                    type: integer
                    example: 12
            price_ranges:
              type: array
              items:
                type: object
                properties:
                  min:
                    type: number
                    example: 100000
                  max:
                    type: number
                    example: 200000
                  count:
                    type: integer
                    example: 25
        search_info:
          type: object
          properties:
            query:
              type: string
              example: "departamento 3 dormitorios"
            total_results:
              type: integer
              example: 150
            search_time:
              type: number
              description: Tiempo de búsqueda en ms
              example: 45.2
...